Cut and paste EffectEditor to new files
