+++ src/effects/Amplify.cpp
@@ -20,6 +20,7 @@
 
 
 #include "Amplify.h"
+#include "EffectEditor.h"
 #include "LoadEffects.h"
 
 #include <math.h>
