+++ src/effects/ChangePitch.cpp
@@ -19,6 +19,7 @@ the pitch without changing the tempo.
 
 #if USE_SOUNDTOUCH
 #include "ChangePitch.h"
+#include "EffectEditor.h"
 #include "LoadEffects.h"
 
 #if USE_SBSMS
