+++ src/effects/ChangeSpeed.cpp
@@ -15,6 +15,7 @@
 
 
 #include "ChangeSpeed.h"
+#include "EffectEditor.h"
 #include "LoadEffects.h"
 
 #include <math.h>
