+++ src/effects/FindClipping.cpp
@@ -17,11 +17,9 @@
 \brief FindClippingDialog used with EffectFindClipping
 
 *//*******************************************************************/
-
-
-
 #include "FindClipping.h"
 #include "AnalysisTracks.h"
+#include "EffectEditor.h"
 #include "LoadEffects.h"
 
 #include <math.h>
