+++ src/effects/vamp/VampEffect.cpp
@@ -10,12 +10,10 @@
   http://www.vamp-plugins.org/
 
 **********************************************************************/
-
-
-
 #if defined(USE_VAMP)
 #include "VampEffect.h"
 #include "../AnalysisTracks.h"
+#include "../EffectEditor.h"
 
 #include <vamp-hostsdk/Plugin.h>
 #include <vamp-hostsdk/PluginChannelAdapter.h>
